feat: generate OpenAPI spec to shared/api for frontend consumption
- Add OpenApiSpecGeneratorTest that exports openapi.json via MockMvc
- Add H2 profile (application-openapi.yml) so spec generation works without Oracle
- Fix Coordinates to use BigDecimal (compatible with both H2 and Oracle)
- Generated spec: 41 operations, all DTOs and schemas

Frontend can generate TypeScript types from shared/api/openapi.json
Usage: ./gradlew test --tests "pl.aerobrains.OpenApiSpecGeneratorTest"

+++ backend/src/main/kotlin/pl/aerobrains/fleet/application/LandingSiteService.kt
@@ -45,7 +45,10 @@ class LandingSiteService(
         }
 
         landingSite.name = request.name
-        landingSite.coordinates = Coordinates(request.latitude, request.longitude)
+        landingSite.coordinates = Coordinates(
+            request.latitude.toBigDecimal(),
+            request.longitude.toBigDecimal()
+        )
 
         return landingSiteMapper.toResponse(landingSiteRepository.save(landingSite))
     }
