+++ backend/src/main/kotlin/pl/aerobrains/fleet/domain/Coordinates.kt
@@ -2,12 +2,13 @@ package pl.aerobrains.fleet.domain
 
 import jakarta.persistence.Column
 import jakarta.persistence.Embeddable
+import java.math.BigDecimal
 
 @Embeddable
 data class Coordinates(
     @Column(nullable = false, precision = 10, scale = 7)
-    val latitude: Double,
+    val latitude: BigDecimal,
 
     @Column(nullable = false, precision = 10, scale = 7)
-    val longitude: Double
+    val longitude: BigDecimal
 )
